Initial commit
Primitive types like int and bool seem to 'just work'. String types mostly work, seem to send to rust fine, and rust can manipulate them, but when rust sends them back the utf8 is malformed, then lua can't print them properly and if we send it back to rust to print we get an error trying to convert the c string to utf8. Perhaps the culprit here is the `(&rust_string).as_ptr() as *const c_char` line?